requirements.txt
ultralytics
torch
torchvision
opencv-python
tqdm
psutil
requests
py-cpuinfo
matplotlib
pyyaml
pycocotools